# Use xmm for stack prolog zeroing rather than rep stos