… src/coreclr/src/jit/codegenxarch.cpp
… src/coreclr/src/jit/compiler.hpp
… src/coreclr/src/jit/emit.cpp
… src/coreclr/src/jit/emit.h
… src/coreclr/src/jit/emitxarch.cpp
… src/coreclr/src/jit/emitxarch.h