@benaadams 2020-02-19
```
Total bytes of diff: 13834 (0.43% of base)
    diff is a regression.

Top file regressions (bytes):
    13834 : System.Private.CoreLib.dasm (0.43% of base)

1 total files with Code Size differences (0 improved, 1 regressed), 0 unchanged.

Top method regressions (bytes):
    516 ( 6.78% of base) : ValueTuple`8:Equals(ValueTuple`8):bool:this (17 methods)
    516 ( 6.78% of base) : ValueTuple`8:CompareTo(ValueTuple`8):int:this (17 methods)
    482 ( 3.01% of base) : ValueTuple`8:ToString():String:this (17 methods)
    482 ( 3.15% of base) : ValueTuple`8:System.IValueTupleInternal.ToStringEnd():String:this (17 methods)
    481 (15.34% of base) : ValueTuple`8:Equals(Object):bool:this (17 methods)
    481 ( 4.70% of base) : ValueTuple`8:System.Collections.IStructuralEquatable.Equals(Object,IEqualityComparer):bool:this (17 methods)
    481 ( 9.53% of base) : ValueTuple`8:System.IComparable.CompareTo(Object):int:this (17 methods)
    481 ( 3.99% of base) : ValueTuple`8:System.Collections.IStructuralComparable.CompareTo(Object,IComparer):int:this (17 methods)
    460 ( 4.23% of base) : EqualityComparer`1:System.Collections.IEqualityComparer.Equals(Object,Object):bool:this (42 methods)
    448 (27.83% of base) : EqualityComparer`1:IndexOf(ref,ValueTuple`8,int,int):int:this (7 methods)
    448 (27.48% of base) : EqualityComparer`1:LastIndexOf(ref,ValueTuple`8,int,int):int:this (7 methods)
    441 ( 4.52% of base) : Comparer`1:System.Collections.IComparer.Compare(Object,Object):int:this (34 methods)
    430 (19.87% of base) : TupleExtensions:CreateLong(__Canon,__Canon,__Canon,__Canon,__Canon,__Canon,__Canon,ValueTuple`8):ValueTuple`8 (7 methods)
    365 ( 1.97% of base) : ValueTuple`8:GetHashCode():int:this (17 methods)
    306 ( 4.48% of base) : TupleExtensions:ToValueTuple(Tuple`8):ValueTuple`8 (14 methods)
    141 ( 4.84% of base) : ValueTuple`8:System.Runtime.CompilerServices.ITuple.get_Item(int):Object:this (17 methods)
    141 ( 2.52% of base) : EqualityComparer`1:System.Collections.IEqualityComparer.GetHashCode(Object):int:this (42 methods)
    136 ( 8.79% of base) : GenericEqualityComparer`1:IndexOf(ref,ValueTuple`8,int,int):int:this (7 methods)
    136 ( 8.68% of base) : GenericEqualityComparer`1:LastIndexOf(ref,ValueTuple`8,int,int):int:this (7 methods)
    130 (13.10% of base) : ValueTuple`8:System.Runtime.CompilerServices.ITuple.get_Length():int:this (17 methods)
    106 ( 9.17% of base) : CustomAttributeData:Init(DllImportAttribute):this
    101 (10.51% of base) : DateTimeParse:TryParseExactMultiple(ReadOnlySpan`1,ref,DateTimeFormatInfo,int,byref):bool (2 methods)
    100 (11.22% of base) : ManyElementAsyncLocalValueMap:Set(IAsyncLocal,Object,bool):IAsyncLocalValueMap:this
     90 ( 4.26% of base) : EventSource:WriteImpl(String,byref,Object,long,long,TraceLoggingEventTypes):this
     89 ( 1.07% of base) : ArraySortHelper`2:IntroSort(Span`1,Span`1,int,IComparer`1) (11 methods)

Top method improvements (bytes):
     -2 (-0.28% of base) : ILStubClass:IL_STUB_PInvoke(SafeRegistryHandle,String,ref,byref,ref,byref):int (2 methods)
     -2 (-0.29% of base) : ILStubClass:IL_STUB_PInvoke(SafeRegistryHandle,String,ref,byref,byref,byref):int (2 methods)
     -2 (-0.05% of base) : Utf8Parser:TryParse(ReadOnlySpan`1,byref,byref,ushort):bool (16 methods)
     -1 (-0.54% of base) : ILStubClass:IL_STUB_PInvoke(byref,byref,byref):bool
     -1 (-0.26% of base) : ILStubClass:IL_STUB_PInvoke(SafeRegistryHandle,int,ref,byref,long,ref,ref,ref):int
     -1 (-0.19% of base) : DateTimeFormat:Format(DateTime,String,IFormatProvider,TimeSpan):String
     -1 (-0.18% of base) : MemoryExtensions:IndexOf(ReadOnlySpan`1,ReadOnlySpan`1,int):int
     -1 (-0.15% of base) : Utf8Formatter:TryFormatFloatingPoint(__Canon,Span`1,byref,StandardFormat):bool
     -1 (-0.15% of base) : ExecutionContext:OnValuesChanged(ExecutionContext,ExecutionContext)
     -1 (-0.19% of base) : Task:WaitAllBlockingCore(List`1,int,CancellationToken):bool
     -1 (-0.23% of base) : NativeOrStaticEventRegistrationImpl:GetEventRegistrationTokenTableInternal(Object,Action`1,byref,bool):ConditionalWeakTable`2
     -1 (-0.34% of base) : FileLoadException:FormatFileLoadExceptionMessage(String,int):String
     -1 (-0.17% of base) : <FinishWriteAsync>d__63:MoveNext():this
     -1 (-0.15% of base) : <DisposeAsyncCore>d__99:MoveNext():this
     -1 (-0.12% of base) : Path:Combine(ref):String
     -1 (-0.42% of base) : PathInternal:RemoveRelativeSegments(String,int):String
     -1 (-0.18% of base) : <DisposeAsyncCore>d__33:MoveNext():this
     -1 (-0.13% of base) : EventSource:EnsureDescriptorsInitialized():this

453 total methods with Code Size differences (18 improved, 435 regressed), 19947 unchanged.
```

@jkotas Went for a higher cut off as there aren't the alignment improvements for larger block is this regression acceptable? Can reduce at a future PR if it does adjustments for alignment.

Also needs this higher value to pick up some methods in Kestrel and 128 bytes is amount mentioned in Intel optimization manual /cc @tannergooding https://github.com/dotnet/runtime/pull/32442#issuecomment-587094492