@EgorBo 2020-02-19
@benaadams let me know if you have samples and want to compare with mono-LLVM which should emit something similar if I am not mistaken.