@benaadams 2020-02-19
Example change from ``ValueTuple`6:Equals(Object):bool:this``

```diff
    sub      rsp, 88
-   mov      rsi, rcx
-   lea      rdi, [rsp+20H]
-   mov      ecx, 12
-   xor      rax, rax
-   rep stosd 
-   mov      rcx, rsi
+   xorps    xmm4, xmm4
+   movdqu   xmmword ptr [rsp+20H], xmm4
+   movdqu   xmmword ptr [rsp+30H], xmm4
+   movdqu   xmmword ptr [rsp+40H], xmm4
    mov      qword ptr [rsp+50H], rdx
```
Worst single method regression `CustomAttributeData:Init(DllImportAttribute):this`
```diff
    sub      rsp, 304
-   mov      rsi, rcx
-   lea      rdi, [rsp+40H]
-   mov      ecx, 60
-   xor      rax, rax
-   rep stosd 
-   mov      rcx, rsi
+   xorps    xmm4, xmm4
+   movdqu   xmmword ptr [rsp+40H], xmm4
+   movdqu   xmmword ptr [rsp+50H], xmm4
+   movdqu   xmmword ptr [rsp+60H], xmm4
+   movdqu   xmmword ptr [rsp+70H], xmm4
+   movdqu   xmmword ptr [rsp+80H], xmm4
+   movdqu   xmmword ptr [rsp+90H], xmm4
+   movdqu   xmmword ptr [rsp+A0H], xmm4
+   movdqu   xmmword ptr [rsp+B0H], xmm4
+   movdqu   xmmword ptr [rsp+C0H], xmm4
+   movdqu   xmmword ptr [rsp+D0H], xmm4
+   movdqu   xmmword ptr [rsp+E0H], xmm4
+   movdqu   xmmword ptr [rsp+F0H], xmm4
+   movdqu   xmmword ptr [rsp+100H], xmm4
+   movdqu   xmmword ptr [rsp+110H], xmm4
+   movdqu   xmmword ptr [rsp+120H], xmm4
    mov      rbp, rcx
```
Though this is still within the range amount of `xmm` zeroing performed by `clang` https://github.com/dotnet/runtime/pull/32442#issuecomment-587554325