@AndyAyersMS 2020-02-19
cc @erozenfeld 

Do we need to go all the way to 256 bytes for methods in Kestrel? 

It would be good to highlight the methods that need a lot of prolog zeroing so we can see if other parts of the heuristic need reworking, or whether other changes -- either to the sources or to the jit -- might be able to remove the need for this much zeroing.

Can you look at code size impact outside of corelib (say via `-f` to `jit-diff`)?