@tannergooding 2020-02-19
> Do we need to go all the way to 256 bytes for methods in Kestrel?

256 bytes looks to be what Clang/LLVM do by default (or 512 bytes for AVX2): https://godbolt.org/z/geQ9u9
160 bytes looks to be what MSVC does by default: https://godbolt.org/z/dphgZN
Both fallback to calling `memset` (rather than using `rep stosb/w/d`) above their cutoffs.

Perhaps this is one of the cases where we could generate the larger code unless the user has explicitly opted for "small code generation"?