@tannergooding 2020-02-19
When optimizing for size; MSVC (`-O1`) looks to drop the cutoff to 64 bytes while Clang (`-Os`) drops the cutoff to 128 bytes (XMM) or 256 bytes (YMM).