@benaadams 2020-02-24
@ahsonkhan effects on Json perf tests

better: 81, geomean: 1.094
worse: 14, geomean: 1.074
total diff: 95

| Slower                                                                           | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Json.Tests.Perf_Ctor.Ctor(Formatted: True, SkipValidation: True)     |      1.37 |            17.99 |            24.62 |         |
| System.Text.Json.Tests.Perf_Guids.WriteGuids(Formatted: False, SkipValidation: T |      1.16 |       3337125.33 |       3869972.00 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSpanEmptyLoop(IsDataCompact: True, TestCa |      1.08 |          3494.72 |          3782.83 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSpanEmptyLoop(IsDataCompact: True, TestCa |      1.06 |          2343.59 |          2493.02 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSingleSpanSequenceEmptyLoop(IsDataCompact |      1.05 |          7970.05 |          8405.27 |         |
| System.Text.Json.Tests.Perf_Reader.ReadReturnBytes(IsDataCompact: True, TestCase |      1.05 |         76498.87 |         79963.71 |         |
| System.Text.Json.Tests.Perf_Reader.ReadReturnBytes(IsDataCompact: True, TestCase |      1.05 |          6399.28 |          6688.48 |         |
| System.Text.Json.Tests.Perf_Strings.WriteStringsUtf16(Formatted: True, SkipValid |      1.04 |      71057200.00 |      74125400.00 |         |
| System.Text.Json.Tests.Perf_Booleans.WriteBooleans(Formatted: True, SkipValidati |      1.04 |       2104204.24 |       2190349.57 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSingleSpanSequenceEmptyLoop(IsDataCompact |      1.04 |          3558.66 |          3702.29 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSingleSpanSequenceEmptyLoop(IsDataCompact |      1.04 |          5341.46 |          5548.98 |         |
| System.Text.Json.Tests.Perf_Booleans.WriteBooleans(Formatted: False, SkipValidat |      1.04 |       1604892.95 |       1663163.75 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSpanEmptyLoop(IsDataCompact: True, TestCa |      1.03 |           711.29 |           733.65 |         |
| System.Text.Json.Tests.Perf_Strings.WriteStringsUtf8(Formatted: False, SkipValid |      1.02 |      56880650.00 |      58238537.50 |         |

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.70 |         11572.86 |          6827.37 |         |
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.55 |          1106.50 |           716.11 |         |
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.53 |          1148.77 |           749.75 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.23 |           238.92 |           194.01 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.23 |           259.48 |           211.49 |         |
| System.Text.Json.Tests.Perf_Deep.WriteDeepUtf16(Formatted: False, SkipValidation |      1.21 |       4840005.88 |       3984675.81 |         |
| System.Text.Json.Tests.Perf_Deep.WriteDeepUtf8(Formatted: False, SkipValidation: |      1.21 |       4842515.38 |       4001904.84 |         |
| System.Text.Json.Tests.Perf_Deep.WriteDeepUtf16(Formatted: False, SkipValidation |      1.20 |       4937916.00 |       4111799.17 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSingleSpanSequenceEmptyLoop(IsDataCompact |      1.18 |           125.96 |           106.78 |         |
| System.Text.Json.Tests.Perf_Deep.WriteDeepUtf8(Formatted: False, SkipValidation: |      1.18 |       4930920.00 |       4185491.53 |         |
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.16 |           137.97 |           119.32 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSingleSpanSequenceEmptyLoop(IsDataCompact |      1.15 |           137.67 |           119.99 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.14 |          4824.78 |          4224.15 |         |
| System.Text.Json.Tests.Perf_Strings.WriteStringsUtf16(Formatted: False, SkipVali |      1.13 |       7611028.13 |       6741727.03 |         |
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.13 |           120.34 |           106.69 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSpanEmptyLoop(IsDataCompact: False, TestC |      1.13 |           112.54 |            99.96 |         |
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.12 |           141.01 |           125.95 |         |
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.12 |           139.43 |           124.67 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.11 |          3647.84 |          3290.39 |         |
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.11 |           138.85 |           125.39 |         |
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.10 |           119.48 |           108.40 |         |
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.09 |         12995.99 |         11900.49 |         |
| System.Text.Json.Tests.Perf_Base64.WriteByteArrayAsBase64_NoEscaping(NumberOfByt |      1.09 |            96.34 |            88.27 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf16(Formatted: False, SkipValidati |      1.09 |           813.75 |           745.92 |         |
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.09 |           120.22 |           110.22 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf16(Formatted: False, SkipValidati |      1.09 |           860.28 |           789.40 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.08 |          1283.97 |          1183.45 |         |
| System.Text.Json.Tests.Perf_Strings.WriteStringsUtf16(Formatted: True, SkipValid |      1.08 |       8215393.33 |       7576784.85 |         |
| System.Text.Json.Tests.Perf_Doubles.WriteDoubles(Formatted: True, SkipValidation |      1.08 |      27176237.50 |      25064325.00 |         |
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.08 |         12843.96 |         11876.66 |         |
| System.Text.Json.Tests.Perf_Strings.WriteStringsUtf16(Formatted: False, SkipVali |      1.08 |      12092662.50 |      11188063.64 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSpanEmptyLoop(IsDataCompact: True, TestCa |      1.08 |           102.65 |            95.00 |         |
| System.Text.Json.Tests.Perf_Strings.WriteStringsUtf16(Formatted: False, SkipVali |      1.08 |      75887800.00 |      70338200.00 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf16(Formatted: True, SkipValidatio |      1.08 |           983.54 |           912.62 |         |
| System.Text.Json.Tests.Perf_Doubles.WriteDoubles(Formatted: True, SkipValidation |      1.08 |      27230387.50 |      25295181.25 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf8(Formatted: False, SkipValidatio |      1.08 |           720.26 |           669.47 |         |
| System.Text.Json.Tests.Perf_Doubles.WriteDoubles(Formatted: False, SkipValidatio |      1.08 |      26360700.00 |      24508177.78 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSingleSpanSequenceEmptyLoop(IsDataCompact |      1.08 |           755.06 |           702.06 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf8(Formatted: False, SkipValidatio |      1.07 |           766.89 |           714.18 |         |
| System.Text.Json.Tests.Perf_Base64.WriteByteArrayAsBase64_HeavyEscaping(NumberOf |      1.07 |            94.14 |            87.68 |         |
| System.Text.Json.Tests.Perf_Reader.ReadReturnBytes(IsDataCompact: False, TestCas |      1.07 |           125.12 |           116.57 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.07 |           983.60 |           917.77 |         |
| System.Text.Json.Tests.Perf_Strings.WriteStringsUtf16(Formatted: True, SkipValid |      1.07 |       8179766.67 |       7632827.27 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.07 |         15569.94 |         14555.17 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf16(Formatted: True, SkipValidatio |      1.07 |          1007.76 |           942.14 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf8(Formatted: True, SkipValidation |      1.07 |           921.71 |           861.72 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.07 |          9835.59 |          9225.88 |         |
| System.Text.Json.Tests.Perf_Doubles.WriteDoubles(Formatted: False, SkipValidatio |      1.06 |      26188868.75 |      24645475.00 |         |
| System.Text.Json.Tests.Perf_Strings.WriteStringsUtf16(Formatted: False, SkipVali |      1.06 |       7626887.88 |       7189622.86 |         |
| System.Text.Json.Tests.Perf_Base64.WriteByteArrayAsBase64_HeavyEscaping(NumberOf |      1.06 |           208.92 |           196.95 |         |
| System.Text.Json.Tests.Perf_Strings.WriteStringsUtf16(Formatted: True, SkipValid |      1.06 |      12747872.22 |      12035400.00 |         |
| System.Text.Json.Tests.Perf_Strings.WriteStringsUtf8(Formatted: False, SkipValid |      1.06 |      61024775.00 |      57838575.00 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSingleSpanSequenceEmptyLoop(IsDataCompact |      1.05 |          8329.30 |          7904.58 |         |
| System.Text.Json.Tests.Perf_Strings.WriteStringsUtf8(Formatted: True, SkipValida |      1.05 |       4693856.60 |       4461803.57 |         |
| System.Text.Json.Tests.Perf_Segment.ReadSingleSegmentSequenceByN(numberOfBytes:  |      1.05 |        992744.18 |        943706.08 |         |
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.05 |           165.94 |           158.08 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf8(Formatted: True, SkipValidation |      1.05 |           879.06 |           838.69 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.05 |          6692.95 |          6386.26 |         |
| System.Text.Json.Tests.Perf_Strings.WriteStringsUtf16(Formatted: False, SkipVali |      1.05 |      11856433.33 |      11325256.82 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSingleSpanSequenceEmptyLoop(IsDataCompact |      1.04 |         94113.68 |         90169.41 |         |
| System.Text.Json.Tests.Perf_Strings.WriteStringsUtf8(Formatted: True, SkipValida |      1.04 |       9264598.15 |       8889771.43 |         |
| System.Text.Json.Tests.Perf_Depth.ReadSpanEmptyLoop(Depth: 1)                    |      1.04 |            97.65 |            93.79 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.04 |         78839.50 |         75770.54 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSingleSpanSequenceEmptyLoop(IsDataCompact |      1.04 |          1331.75 |          1280.70 |         |
| System.Text.Json.Tests.Perf_Strings.WriteStringsUtf8(Formatted: True, SkipValida |      1.04 |       4685367.92 |       4514592.73 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.04 |          4425.80 |          4265.39 |         |
| System.Text.Json.Tests.Perf_Reader.ReadReturnBytes(IsDataCompact: False, TestCas |      1.03 |         14186.85 |         13723.27 |         |
| System.Text.Json.Tests.Perf_Reader.ReadReturnBytes(IsDataCompact: True, TestCase |      1.03 |          2784.09 |          2696.61 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.03 |        106975.87 |        103708.50 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSpanEmptyLoop(IsDataCompact: False, TestC |      1.03 |          3436.04 |          3331.55 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.03 |         11265.65 |         10928.63 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSpanEmptyLoop(IsDataCompact: True, TestCa |      1.03 |          5449.46 |          5291.53 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.03 |          2314.19 |          2249.15 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf16(Formatted: False, SkipValidati |      1.03 |       1622613.73 |       1577841.14 |         |
| System.Text.Json.Tests.Perf_Reader.ReadMultiSpanSequenceEmptyLoop(IsDataCompact: |      1.03 |         10181.32 |          9902.96 |         |
| System.Text.Json.Tests.Perf_Reader.ReadSpanEmptyLoop(IsDataCompact: True, TestCa |      1.03 |         68491.36 |         66709.24 |         |
| System.Text.Json.Tests.Perf_Strings.WriteStringsUtf8(Formatted: True, SkipValida |      1.03 |       9096440.74 |       8860973.21 |         |
| System.Text.Json.Tests.Perf_Segment.ReadSingleSegmentSequence(TestCase: Json4KB) |      1.03 |          8234.92 |          8027.36 |         |
| System.Text.Json.Tests.Utf8JsonReaderCommentsTests.Utf8JsonReaderCommentParsing( |      1.02 |           160.27 |           156.49 |         |
| System.Text.Json.Tests.Perf_Strings.WriteStringsUtf8(Formatted: True, SkipValida |      1.02 |      58887750.00 |      57558900.00 |         |
| System.Text.Json.Tests.Perf_Reader.ReadReturnBytes(IsDataCompact: True, TestCase |      1.02 |          9835.37 |          9623.42 |         |